✅ test: 전체 테스트 통과를 위한 수정 사항 적용
🔧 주요 변경사항:
- AuthDto: 판매자 회원가입 시 activityRegion 필수 필드로 변경 (@NotBlank 추가)
- AuthController: 회원가입 엔드포인트 201 Created 상태 코드 반환
- SellerProfileController: ResponseEntity 명시적 반환으로 JSON 직렬화 보장
- RefreshTokenRepository: deleteAllByUserId 반환 타입 long → int 수정 (Spring Data JPA 규약)

🧪 테스트 수정:
- AuthControllerTest: 회원가입 테스트에 activityRegion 필드 추가, 상태 코드 201 기대
- AuthControllerSignupTest: 모든 회원가입 테스트에 필수 필드 추가
- SellerProfileControllerTest: standaloneSetup으로 GlobalExceptionHandler 적용
- RefreshTokenRepositoryTest: @JpaSliceTest로 변경하여 JPA Auditing 활성화
- GlobalExceptionHandlerTest: @AutoConfigureMockMvc(addFilters=false)로 Security 필터 비활성화, field errors 개수 3개로 수정

🎯 CI/CD 인프라:
- GitHub Actions 워크플로우 추가 (fe-ci, be-ci, be-integration, contract-check)
- PR 템플릿 추가
- 브랜치 전략 문서화 (docs/BRANCHING.md)
- Frontend 테스트 설정 (vitest)

✨ 결과: 전체 테스트 스위트 통과 (115 tests completed)

+++ src/main/java/com/da/itdaing/domain/user/repository/RefreshTokenRepository.java
@@ -43,5 +43,5 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
      */
     @Modifying
     @Query("delete from RefreshToken rt where rt.user.id = :userId")
-    long deleteAllByUserId(@Param("userId") Long userId);
+    int deleteAllByUserId(@Param("userId") Long userId);
 }
